Move extension `openUrlInChromeCustomTab` to :androidutils module.
